+++ libsqsh/src/posix/file_ext.c
@@ -181,7 +181,7 @@ file_iterator_mt(
 		goto out;
 	}
 
-	size_t block_offset = 0;
+	uint64_t block_offset = 0;
 	for (sqsh_index_t i = 0; i < block_count; i++) {
 		mt->blocks[i].mt = mt;
 		mt->blocks[i].block_offset = block_offset;
